Continued arrangement with arraylist started: with more errors xdxd

+++ src/Debug.java
@@ -5,7 +5,7 @@ public class Debug extends Processing {
     public static void main(String[] args) {
 
     }
-    public static void printArray(long[] input){
+    public static void printArray(int[] input){
         for(int i = 0; i< input.length; i++){
             System.out.print(input[i]);
         }
